Add Directories and Files
Added three directories to separate content by language. HTML, CSS and Javascript containing their related files.

Created one file in each
- index.html: Homepage
- style.css: Styling
- script.js: Scripting

Added boilerplate code to the index.html to construct the bare bones and gave it a title as well as the relevant script and link to relate it to the relevant files.